Split LPA context
LPA context split and RequestActivationKey was created. Added methods needed for some new feature tests
