+++ service-front/app/features/context/UI/LpaContext.php
@@ -25,7 +25,6 @@ use Psr\Http\Message\RequestInterface;
  * @property string $userFirstName
  * @property int    $userId
  * @property string $userSurname
- * @property string $activationCode
  * @property string $codeCreatedDate
  */
 class LpaContext implements Context
@@ -139,43 +138,11 @@ class LpaContext implements Context
     }
 
     /**
-     * @Then /^a letter is requested containing a one time use code$/
-     */
-    public function aLetterIsRequestedContainingAOneTimeUseCode()
-    {
-        $this->ui->assertPageAddress('/lpa/check-answers');
-
-        $this->ui->assertElementContainsText('h1', 'We\'re posting you an activation key');
-    }
-
-    /**
      * @Given /^an attorney can be removed from acting on a particular LPA$/
      */
     public function anAttorneyCanBeRemovedFromActingOnAParticularLpa()
     {
         // Not needed for this context
-    }
-
-    /**
-     * @Given I already have a valid activation key for my LPA
-     */
-    public function iAlreadyHaveAValidActivationKeyForMyLpa()
-    {
-        $this->activationCode = 'ACTVATIONCOD';
-        $this->codeCreatedDate = (new DateTime())->modify('-15 days')->format('Y-m-d');
-    }
-
-    /**
-     * @Then /^I am asked to check my answers before requesting an activation key$/
-     */
-    public function iAmAskedToCheckMyAnswersBeforeRequestingAnActivationKey()
-    {
-        $this->ui->assertPageAddress('/lpa/check-answers');
-        $this->ui->assertPageContainsText('Check your answers');
-        $this->ui->assertPageContainsText('700000000001');
-        $this->ui->assertPageContainsText('The Attorney Person');
-        $this->ui->assertPageContainsText('9 February 1998');
-        $this->ui->assertPageContainsText('ABC123');
     }
 
     /**
@@ -194,16 +161,6 @@ class LpaContext implements Context
     public function iAmInactiveAgainstTheLpaOnMyAccount()
     {
         $this->lpaData['actor']['details']['systemStatus'] = false;
-    }
-
-    /**
-     * @Then I am informed that an LPA could not be found with these details
-     */
-    public function iAmInformedThatAnLPACouldNotBeFoundWithTheseDetails()
-    {
-        $this->ui->assertPageAddress('/lpa/check-answers');
-
-        $this->ui->assertElementContainsText('h1', 'We could not find an LPA with the details you entered');
     }
 
     /**
@@ -318,15 +275,8 @@ class LpaContext implements Context
         $this->iAmNavigatedToTheInstructionsAndPreferencesPage();
     }
 
-    /**
-     * @Given /^I am on the request an activation key page$/
-     * @Given /^I am on the add an older LPA page$/
-     */
-    public function iAmOnTheRequestAnActivationKeyPage()
-    {
-        $this->ui->visit('/lpa/add-by-paper');
-        $this->ui->assertPageAddress('/lpa/add-by-paper');
-    }
+
+
 
     /**
      * @Then /^I am shown a not found error$/
@@ -334,21 +284,6 @@ class LpaContext implements Context
     public function iAmShownANotFoundError()
     {
         $this->ui->assertResponseStatus(404);
-    }
-
-    /**
-     * @Then /^I am taken back to previous page where I can see my answers and change them$/
-     */
-    public function iAmTakenBackToPreviousPageWhereICanSeeMyAnswersAndChangeThem()
-    {
-        $this->ui->assertPageAddress('/lpa/add-by-paper');
-        $this->ui->assertFieldContains('opg_reference_number', '700000000001');
-        $this->ui->assertFieldContains('first_names', 'The Attorney');
-        $this->ui->assertFieldContains('last_name', 'Person');
-        $this->ui->assertFieldContains('dob[day]', '09');
-        $this->ui->assertFieldContains('dob[month]', '02');
-        $this->ui->assertFieldContains('dob[year]', '1998');
-        $this->ui->assertFieldContains('postcode', 'ABC123');
     }
 
     /**
@@ -384,38 +319,6 @@ class LpaContext implements Context
     {
         $this->lpaData['actor']['type'] = 'donor';
         unset($this->lpaData['actor']['details']['systemStatus']);
-    }
-
-    /**
-     * @Then I am told that I cannot request an activation key
-     */
-    public function iAmToldThatICannotRequestAnActivationKey()
-    {
-        $this->ui->assertPageAddress('/lpa/check-answers');
-
-        $this->ui->assertElementContainsText('h1', 'We cannot send an activation key for that LPA');
-
-        $dashboardButton = $this->ui->getSession()->getPage()->findAll(
-            'css',
-            'main a.govuk-button[href^="/lpa/dashboard"]'
-        );
-        if (count($dashboardButton) !== 1) {
-            throw new AssertionFailedError('Did not find button to navigate to dashboard');
-        }
-
-        $logoutButton = $this->ui->getSession()->getPage()->findAll('css', 'main a.govuk-button[href^="/logout"]');
-        if (count($logoutButton) !== 1) {
-            throw new AssertionFailedError('Did not find button to logout');
-        }
-    }
-
-    /**
-     * @Then I am told that I have an activation key for this LPA and where to find it
-     */
-    public function iAmToldThatIHaveAnActivationKeyForThisLpaAndWhereToFindIt()
-    {
-        $this->ui->assertPageAddress('/lpa/check-answers');
-        $this->ui->assertElementContainsText('h1', 'We\'ve already sent you an activation key for this LPA');
     }
 
     /**
@@ -1265,15 +1168,6 @@ class LpaContext implements Context
     }
 
     /**
-     * @When /^I confirm that those details are correct$/
-     */
-    public function iConfirmThatThoseDetailsAreCorrect()
-    {
-        $this->ui->assertPageAddress('/lpa/check-answers');
-        $this->ui->pressButton('Continue');
-    }
-
-    /**
      * @When /^I do not confirm cancellation of the chosen viewer code$/
      */
     public function iDoNotConfirmCancellationOfTheChosenViewerCode()
@@ -1407,7 +1301,6 @@ class LpaContext implements Context
     /**
      * @Given I have been given access to use an LPA via credentials
      * @Given I have added an LPA to my account
-     * @Given I have been given access to use an LPA via a paper document
      */
     public function iHaveBeenGivenAccessToUseAnLPAViaCredentials()
     {
@@ -1511,16 +1404,6 @@ class LpaContext implements Context
     }
 
     /**
-     * @Given /^I have requested an activation key with valid details$/
-     */
-    public function iHaveRequestedAnActivationKeyWithValidDetails()
-    {
-        $this->iAmOnTheRequestAnActivationKeyPage();
-        $this->iRequestAnActivationKeyWithValidDetails();
-        $this->iAmAskedToCheckMyAnswersBeforeRequestingAnActivationKey();
-    }
-
-    /**
      * @When /^I have shared the access code with organisations to view my LPA$/
      */
     public function iHaveSharedTheAccessCodeWithOrganisationsToViewMyLPA()
@@ -1577,153 +1460,6 @@ class LpaContext implements Context
             );
 
         $this->ui->visit('lpa/view-lpa?lpa=' . $this->userLpaActorToken);
-    }
-
-    /**
-     * @When I provide details from an LPA registered before Sept 2019
-     */
-    public function iProvideDetailsFromAnLpaRegisteredBeforeSept2019()
-    {
-        $this->fillAndSubmitOlderLpaForm();
-
-        // Setup fixture for success response
-        $this->apiFixtures->patch('/v1/lpas/request-letter')
-            ->respondWith(
-                new Response(
-                    StatusCodeInterface::STATUS_BAD_REQUEST,
-                    [],
-                    json_encode(
-                        [
-                            'title' => 'LPA not eligible due to registration date',
-                            'details' => 'LPA not eligible due to registration date',
-                            'data' => [],
-                        ]
-                    )
-                )
-            );
-    }
-
-    /**
-     * @When I provide details that do not match a valid paper document
-     */
-    public function iProvideDetailsThatDoNotMatchAValidPaperDocument()
-    {
-        $this->fillAndSubmitOlderLpaForm();
-
-        // Setup fixture for success response
-        $this->apiFixtures->patch('/v1/lpas/request-letter')
-            ->respondWith(
-                new Response(
-                    StatusCodeInterface::STATUS_BAD_REQUEST,
-                    [],
-                    json_encode(
-                        [
-                            'title' => 'LPA details do not match',
-                            'details' => 'LPA details do not match',
-                            'data' => [],
-                        ]
-                    )
-                )
-            );
-    }
-
-    /**
-     * @When I provide the details from a valid paper document
-     */
-    public function iProvideTheDetailsFromAValidPaperDocument()
-    {
-        $this->fillAndSubmitOlderLpaForm();
-
-        /**
-         * This step definition needs to behave differently dependant on some prior context step
-         */
-        if ($this->activationCode === null) {
-            // Setup fixture for success response
-            $this->apiFixtures->patch('/v1/lpas/request-letter')
-                ->respondWith(
-                    new Response(
-                        StatusCodeInterface::STATUS_NO_CONTENT,
-                        [],
-                        ''
-                    )
-                );
-        } else {
-            // Setup fixture for activation key already existing
-            $this->apiFixtures->patch('/v1/lpas/request-letter')
-                ->respondWith(
-                    new Response(
-                        StatusCodeInterface::STATUS_BAD_REQUEST,
-                        [],
-                        json_encode(
-                            [
-                                'title' => 'Bad request',
-                                'details' => 'LPA has an activation key already',
-                                'data' => [
-                                    'donor_name' => [
-                                        ($this->lpaData['lpa'])->donor->firstname,
-                                        ($this->lpaData['lpa'])->donor->middlenames,
-                                        ($this->lpaData['lpa'])->donor->surname
-                                    ],
-                                    'lpa_type'   => $this->lpaData['lpa']->caseSubtype
-                                ],
-                            ]
-                        )
-                    )
-                );
-        }
-        // API call for Notify
-        $this->apiFixtures->post(Client::PATH_NOTIFICATION_SEND_EMAIL)
-            ->respondWith(new Response(StatusCodeInterface::STATUS_OK, [], json_encode([])))
-            ->inspectRequest(
-                function (RequestInterface $request) {
-                    $params = json_decode($request->getBody()->getContents(), true);
-
-                    assertInternalType('array', $params);
-                    assertArrayHasKey('template_id', $params);
-                    assertArrayHasKey('personalisation', $params);
-                }
-            );
-    }
-
-    /**
-     * @When /^I request an activation key with an invalid DOB format of "([^"]*)" "([^"]*)" "([^"]*)"$/
-     */
-    public function iRequestAnActivationKeyWithAnInvalidDOBFormatOf($day, $month, $year)
-    {
-        $this->ui->assertPageAddress('/lpa/add-by-paper');
-        $this->ui->fillField('opg_reference_number', '700000000001');
-        $this->ui->fillField('first_names', 'Attorney');
-        $this->ui->fillField('last_name', 'Person');
-        $this->ui->fillField('postcode', 'ABC123');
-        $this->ui->fillField('dob[day]', $day);
-        $this->ui->fillField('dob[month]', $month);
-        $this->ui->fillField('dob[year]', $year);
-        $this->ui->pressButton('Continue');
-    }
-
-    /**
-     * @When /^I request an activation key with an invalid lpa reference number format of "([^"]*)"$/
-     */
-    public function iRequestAnActivationKeyWithAnInvalidLpaReferenceNumberFormatOf($referenceNumber)
-    {
-        $this->ui->fillField('opg_reference_number', $referenceNumber);
-        $this->ui->pressButton('Continue');
-    }
-
-    /**
-     * @When /^I request an activation key with valid details$/
-     */
-    public function iRequestAnActivationKeyWithValidDetails()
-    {
-        $this->ui->assertPageAddress('/lpa/add-by-paper');
-        $this->ui->fillField('opg_reference_number', '700000000001');
-        $this->ui->fillField('first_names', 'The Attorney');
-        $this->ui->fillField('last_name', 'Person');
-        $this->ui->fillField('postcode', 'ABC123');
-        $this->ui->fillField('dob[day]', '09');
-        $this->ui->fillField('dob[month]', '02');
-        $this->ui->fillField('dob[year]', '1998');
-        $this->ui->pressButton('Continue');
     }
 
     /**
@@ -2052,13 +1788,7 @@ class LpaContext implements Context
         $this->ui->pressButton('Continue');
     }
 
-    /**
-     * @When /^I request to go back and change my answers$/
-     */
-    public function iRequestToGoBackAndChangeMyAnswers()
-    {
-        $this->ui->clickLink('Change');
-    }
+
 
     /**
      * @Given /^I request to go back and try again$/
@@ -2087,7 +1817,6 @@ class LpaContext implements Context
         $this->lpa->status = $status;
 
         if ($status === 'Revoked') {
-
             // API call for get LpaById
             $this->apiFixtures->get('/v1/lpas/' . $this->userLpaActorToken)
                 ->respondWith(
@@ -2522,25 +2251,6 @@ class LpaContext implements Context
         $this->ui->assertPageContainsText('Health and welfare');
     }
 
-    protected function fillAndSubmitOlderLpaForm()
-    {
-        $this->ui->fillField('opg_reference_number', $this->lpa->uId);
-        $this->ui->fillField(
-            'first_names',
-            $this->lpa->donor->firstname . ' ' . $this->lpa->donor->middlenames
-        );
-        $this->ui->fillField('last_name', $this->lpa->donor->surname);
-
-        $date = new DateTime($this->lpa->donor->dob);
-        $this->ui->fillField('dob[day]', $date->format('d'));
-        $this->ui->fillField('dob[month]', $date->format('m'));
-        $this->ui->fillField('dob[year]', $date->format('Y'));
-
-        $this->ui->fillField('postcode', ($this->lpa->donor->addresses[0])->postcode);
-
-        $this->ui->pressButton('Continue');
-    }
-
     /**
      * @When /^I check access codes of the status changed LPA$/
      */
@@ -2611,28 +2321,6 @@ class LpaContext implements Context
     public function iLostTheLetterReceivedHavingTheActivationKey()
     {
         // Not needed for this context
-    }
-
-    /**
-     * @Then /^I should have an option to regenerate an activation key for the old LPA I want to add$/
-     */
-    public function iShouldHaveAnOptionToRegenerateAnActivationKeyForTheOldLPAIWantToAdd()
-    {
-        $this->iProvideTheDetailsFromAValidPaperDocument();
-        $this->iConfirmThatThoseDetailsAreCorrect();
-        $this->iAmToldThatIHaveAnActivationKeyForThisLpaAndWhereToFindIt();
-
-        $this->ui->assertPageAddress('/lpa/check-answers');
-        $this->ui->assertPageContainsText('Continue and ask for a new key');
-    }
-
-    /**
-     * @Then /^I request for a new activation key again$/
-     */
-    public function iRequestForANewActivationKeyAgain()
-    {
-        $this->iShouldHaveAnOptionToRegenerateAnActivationKeyForTheOldLPAIWantToAdd();
-        $this->ui->pressButton('Continue and ask for a new key');
     }
 
     /**
